requirements/dev.txt
-r base.txt

# Testing
pytest
pytest-django
pytest-cov